+++ sentry/src/main/java/io/sentry/JsonSerializer.java
@@ -72,6 +72,7 @@ public final class JsonSerializer implements ISerializer {
   /**
    * All our custom deserializers need to be registered to be used with the deserializer instance. *
    */
+  @SuppressWarnings("deprecation")
   public JsonSerializer(@NotNull SentryOptions options) {
     this.options = options;